backend/requirements.txt
django==5.1.14
djangorestframework==3.15.2
gunicorn==23.0.0
python-keycloak==4.7.3
PyJWT==2.10.1
cryptography==44.0.2
finnhub-python==2.4.22
django-cors-headers==4.7.0
requests==2.32.3
